requirements.txt
flask
flask_login
flask_cors
flask_sqlalchemy
python-dotenv
psycopg2
gunicorn